requirements.txt
scapy>=2.5.0
cryptography>=41.0.0
pycryptodome>=3.19.0
pytest>=7.4.0
iperf3>=0.1.11 